requirements.txt
# requirements.txt - Personal Strategic Intelligence System

# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-session==0.5.0
flask-socketio>=5.3.0
gunicorn==21.2.0

# Authentication
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
requests-oauthlib==1.3.1

# Multi-Database Storage Architecture
psycopg2-binary>=2.9.9
pgvector>=0.2.4
chromadb>=0.4.22
neo4j>=5.15.0
redis>=5.0.1
boto3>=1.29.0
minio>=7.2.0

# AI & Language Models
anthropic>=0.40.0
openai>=1.10.0
tiktoken>=0.5.0
langchain>=0.1.0
sentence-transformers>=2.2.2

# Web Intelligence & Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.0
httpx>=0.25.0
scrapy>=2.11.0
selenium>=4.15.0

# Data Analysis & Machine Learning
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
networkx>=3.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async & Real-time
asyncio>=3.4.3
websockets>=11.0.0
eventlet>=0.33.0
celery>=5.3.0
kombu>=5.3.0

# Data Processing
python-dateutil==2.8.2
pytz>=2023.3
textblob>=0.17.1
spacy>=3.7.0
nltk>=3.8.1

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk>=1.39.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Existing dependencies
sqlalchemy==2.0.23 